requirements.txt
altair==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
bgstools==0.2.1 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
cycler==0.11.0 ; python_version >= "3.11" and python_version < "4.0"
ffmpeg-python==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
fonttools==4.42.0 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2023.6.0 ; python_version >= "3.11" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.11" and python_version < "4.0"
gitdb==4.0.10 ; python_version >= "3.11" and python_version < "4.0"
gitpython==3.1.32 ; python_version >= "3.11" and python_version < "4.0"
h3==3.7.6 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2023.7.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.19.0 ; python_version >= "3.11" and python_version < "4.0"
kiwisolver==1.4.4 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
matplotlib==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.25.2 ; python_version >= "3.11" and python_version < "4.0"
opencv-python==4.8.0.76 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.24.0 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==12.0.1 ; python_version >= "3.11" and python_version < "4.0"
pydeck==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.16.1 ; python_version >= "3.11" and python_version < "4.0"
pympler==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.11" and python_version < "4.0"
pyproj==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.30.2 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rich==13.5.2 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.9.2 ; python_version >= "3.11" and python_version < "4.0"
shapely==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
smmap==5.0.0 ; python_version >= "3.11" and python_version < "4.0"
st-files-connection @ git+https://github.com/streamlit/files-connection@main ; python_version >= "3.11" and python_version < "4.0"
streamlit==1.25.0 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.2.3 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.3.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
tzlocal==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
validators==0.21.2 ; python_version >= "3.11" and python_version < "4.0"
watchdog==3.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Darwin"
zipp==3.16.2 ; python_version >= "3.11" and python_version < "4.0"
